Extracting components from SigningKeysTable, StreamSessionsTable, StreamsTable, WebhookTable
